Move electrical component enums to their own module
